Cleanup email_integrations + removed unused inntegrations
