+++ api/app/signals/settings/testing.py
@@ -14,22 +14,6 @@ SIGNALS_AUTHZ = {
     'ALWAYS_OK': True,
     'USER_ID_FIELDS': 'sub,email'.split(',')
 }
-# Email integration settings
-EMAIL_INTEGRATIONS = dict(
-    FLEX_HORECA=dict(
-        RECIPIENT_LIST=['test@test.com', ],
-        APPLICABLE_RULES=dict(
-            WEEKDAYS='5,6,7',  # fri, sat, sun
-            END_TIME='04:00',  # 04:00 o'clock
-        )
-    ),
-    TOEZICHT_OR_NIEUW_WEST=dict(
-        RECIPIENT_LIST=['test@test.com', ],
-    ),
-    VTH_NIEUW_WEST=dict(
-        RECIPIENT_LIST=['test@test.com', ],
-    ),
-)
 
 FEATURE_FLAGS['API_SEARCH_ENABLED'] = False  # noqa F405
 FEATURE_FLAGS['SEARCH_BUILD_INDEX'] = False  # noqa F405
